Nutupane - Re-factoring and clean-up, addition of SystemGroup loading.
This re-factors the Nutupane factory to be a re-usable object factory.
The change requires each object using the Nutupane factory to instantiate
an instance of it through `new Nutupane()`.

The current_organization value was re-factored to follow naming
conventions and is now CurrentOrganization.

Includes addition of SystemGroup search functionality in API layer.