965899 - do not include component view label in composite view path
Prior to this change, when a composite view is promoted
(e.g. Library to dev), the content path on pulp file
system would include both name of the composite view
as well as the name of the component view.  The component
view name is not necessary; therefore, it is being removed
from the path.

E.g.
Before:
/var/lib/pulp/published/https/repos/ACME_Corporation/dev/comp_view/zoo_view/custom/zoo_product/zoo_repo

After:
/var/lib/pulp/published/https/repos/ACME_Corporation/dev/comp_view/custom/zoo_product/zoo_repo