Bastion - Adding test infrastructure for engine JavaScript via Karma.
The Karma test runner is built and designed for testing cross-browser
JavaScript and making integration with AngularJS easier. This adds
a Karma configuration file and includes NPM package and Bower component
definitions to load up components in a predictable location for testing.
This is necessary in order for the testing infrastructure to predictably
locate the necessary assets. The tests can be run via 'karma start' and
'karma run'. A gem based test runner such as 'jasmine-rails' was not used
due to restrictions imposed by it's usage and tight coupling with the asset
pipeline and Rails.