918093 - update ES index for packages and other document types
When indexing packages, do not index the 'changelog', 'files' and
'filelist' that is retrieved from pulp.  These fields were already
being omitted when packages were indexed as part of a repo
promotion; however, when individual packages were selected they
were not omitted.

Also, for all object types (e.g. errata, package, package group..etc),
split the creation of the index and the importing of objects in
to 2 separate actions.  The reason for this change is that with
the current version of elastic search, attempting to create an
index that already exists, will generate an error.  Previously,
when that error occured, it would skip the import that followed.