970644 - Allows passing organization name with a dot in the path to providers API.
The default Rails routing returns 404s when dynamic URL segements
contain a dot. This fix adds a constraint to the providers list route
to allow Rails to pass through the organization_id and allow the
controller to handle processing. This is required if an organization
with a dot in the name is created and a user wants to list the providers
from that organization.