960695 - Fixes alignment issue with the attach and remove buttons.
The attach key and remove key button from the available and attached
subscriptions tabs were in different places. This fix standardizes
the actions to the upper right hand corner and the filter box to the
upper left hand corner of the table. Moving the actions to the upper
right is to account for potentially long lists that could hide the
actions from the users view.