Bastion: Adding sticky headers and table body scrolling.
The browser does not make scrolling the body of a table easy when
the table is using responsive based styling. The strategy used here
provides two table heads, one of which is treated as a shadow table head
and the other as the true table head. The shadow table head must exist
to get the styling and centering of table headers with the data correct.
A directive is provided to handle dealing with browser resizes to set
the width and height properly.