ensuring library content is accessible by consumers
required updating the environment content in candlepin
and generating metadata in pulp

previously when refreshing a new CVV would be created and the old
one would be removed from the environment.  This would result in
the candlepin environment being deleted and the repos being deleted form
library as well as any repo bindings being lost in pulp.

This change makes it so that the content view version is created and
all repos are moved to it before the old version is removed from library.
Then when refresh_repos is called instead of operating on the repos within
the old version (that has been deleted), we operate on the versions within
the content view version itself