975061 - prevent deletion of CV from environment with systems
If we delete a content view from an environment the candlepin
environment gets deleted.  This results in all the candlepin consumers
being deleted.   Prevent this from happening.  While this is not the ideal
way, the promotion of content views should be changed soon and we can
address this better there.