978001 - add 'display_name' to rhsm environment listing
When a system registers to katello using subscription manager,
the environment 'label' is used.  This is also what is shown
if a user lists the environments.

There are, however, cases where rhsm tools desire access
to the environment name.  Since we need to support the
existing interface where name=env.label, this commit will
add display_name=env.name to the rhsm api for environments.