issue 2353: system/system group - update backend to support 'update all' packages
This commit contains a minor change to allows support for
an 'update all packages' for a system or system group.
This would be the equivalent to an admin executing
a 'yum update' each system.

This capability was available with pulpv1; however, when
we migrated to pulpv2, we missed the modifications to enable
the same functionality.