Tests: Refactor pulp glue tests to be more independent of one another.
Previously the glue layer tests were all connected through the use
of support functions with their own cassettes and large collections
of tests tied to single cassettes. This introduces the use of a single
cassette per suite of tests and the removal of dedicated support cassettes.
Each suite of tests, a suite being defined by a TestCase class (not a file),
is reponsible for it's own HTTP requests no matter what entity generates them.
This should reduce the amount of cassette re-generation as updates to Pulp
are made or functionality is added to specific entities.