Test: Changing to use a dedicated feed url across all systems that may run the tests.
Previously the feed URL used in tests was dependent upon the system
where the tests were originally ran and disallowed body matching for
repository creation. The feed URL has now been set to '/var/www' and
functionality added to copy the test_repo to this location on every
test run in :all mode. This will allow changes to be made to the repo
and for other developers running the tests to automatically get those
changes. Further, users will no longer need to open up their home
directory to Apache in order to run the tests against a live Pulp.