Bastion: Clean-up and re-organization to get assets precompile to work.
The target for Bastion is to provide a single SCSS/CSS file and a single
JS file in production. The changes herein re-organize some assets and
provide the configuration necessary to generate and precompile the JS
and SCSS files appropriately for Bastion. This commit also includes some
missing module includes and removes some unnecessary ones.