Engine: User: update to assume user is coming from core (e.g. Foreman)
This commit contains some initial changes to assume that the
"User" that is used by the engine comes from ::User (e.g. core, foreman).
In addition, these changes will just extend the behaviors
of that user based on the current behavior of the engine.  Eventually,
we may want to leverage the foreman settings capabilities for
some of those.