Bunch of migration fixes
To get headpin migrate from 1.2 -> 1.3

The existing migration scripts failed when upgrading from headpin 1.2->
1.3 .. This is mostly because the migration scripts use models (to
update services like candlepin & pulp) , which have changed drastically
since the migrations were written.

This commit has the following changes.

1) db/migrate/20130215063241_populate_content_view_environments.rb
Needed to change the way the cp environment label was determined.
The method name to generate a cp_environment_label and id has changed
in the ContentView class to generate_cp_environment_label and id...
So this change needed to be made

2)
db/migrate/20130318171849_migrate_environment_default_content_view_to_version.rb
This file had an issue obtaining default content view associated to an
environment. This migration originally planned to achieve the following
change.
Prior to running this migration each environment pointed to a
default_content_view and the view had its own version. This migration
intends to create a org default view and then make each environment's
content view version point to this org default view.
This change is only relevant for katello, since we know that for headpin
there is no longer any Non-Library environment. So this migration script
tries to scope out the ContentView creation and other parts to only
Katello.

3)
db/migrate/20130515153703_headpin_migration.rb ->
db/migrate/20130715153703_headpin_migration.rb

Moved the headpin migration to be run after the environment_products
removal change happened. This is because this file points to model code
that is drastically different now. When originally run this migration
showed up an error along the lines of "no such column environment_id in
repositories". That is because the migration to add environment_id to
repositories happened much later. Now instead of tinkering and monkey
patching the glue code, it seemed more prudent to move this migration
after the migration that adds environment_id to repositories.