allowing admin permission to be recreated
The Role.make_super_admin_role is designed to create the admin role
with its permissions if it did not exist.  It does not handle the case
that the Role exists but the permission does not.  If minitests are run
before spec tests, this situation occurs and Role.make_super_admin_role will fail
as the Role is locked and no new permissions can be added.