puppet - update content search to include puppet module counts
This commit contains changes to show the puppet modules count
when performing search by content view, product and repository.

Since adding puppet modules adds more information to the
cell on hover, condensing the format to be similar to:
	X / Y / Z

With a (i) tipsy on the far right that allows user to see
it displayed as:
	Packages (X)
	Errata (Y)
	Puppet Modules (Z)

This approach was taken, since the cell would have had to have
been increase in size significantly to include the new attribute.