953656 - Removed all 'disable_with' references
Rails haml form builder provides an attribute called disable_with
that lets one to set a 'disabled' text on the butten when a form is
submitted. For example a button with  "Create" can be changed to
"Creating" when clicked on. However this automagical attribute doesnot
go well with ajax based calls. So this caused issues like bz 953656,
where validation failures wouldnt change "Creating" to "Create" if the
submit failed.
We are automatically disabling the button anyway without changing
the text on a form submit even without the disable_with attribute.
So removed them from all haml files. So the new behaviour will be that
"Create" button will still be "Create" but disabled instead of
"Creating...".