994917: Content Search: Errata: show null severity as ''
There is currently a bug in pulp
(https://bugzilla.redhat.com/show_bug.cgi?id=1009966) which
is resulting all errata not having a severity.  As a result,
when content search shows the errata listing it shows as 'null'.
This commit will simply render a '' (instead of null) when
there is no errata.