Code to help with katello-upgrade and reindexing
Had to update the legacy promotion code to work with new runcible changes.
Also made a slight tweak to the rake reindex call to reindex only
if there was content to reindex. This was done to avoid the following
error one gets when running katello-upgrade

[ERROR] Too many exceptions occured, giving up. The HTTP response was:
400 > {"error":"Failed to derive xcontent from
org.elasticsearch.common.bytes.BytesArray@0"}

Also added a print statement to print what is getting reindexed.