Bug 1013681: Re-enables after sync notifications when Pulp sends a sync completed notification and updates the tooltip rendering on the sync management page.
Previously we disabled task updates and notifications for sync complete
notifications from Pulp due to a missing task_id parameter. This fix
re-enables and updates the code to show the newly formatted error details
to the user. This also updates the tooltip shown on the sync management
page to give users a better error output. Further, the error icon is also
clickable and takes users to the notifications page.