Bug 1011528: Fixing broken default environment by simplifying to the choice down to a default organization and preferring library registration when no environment is given.
When content views were added, the default environment functionality
broke from the UI perspective since a user would need to select both
an environment and a content view. To simplify this interaction the
default environment selection was removed and replaced by choosing
just a default organization. This default organization is also
the user's favorite organization and will result in the user being
logged into that organization by default.