1023525 - fixing new provider page on repo discovery
Previously the new provider state on repodiscovery
was trying to re-use the existing state, however this
results in none of the items being saved properly, and
all user selection is lost after creating the provider.

This sets up a child state so that all user selection is
saved, as well as ensures that the provider is selected
after the user returns to the repo discovery page.