Bug 1024055: Changes content view definition puppet repo selector to basic HTMl select to prevent choices from being hidden from the user.
When adding content to a content view definition, the yum products and
repositories can push the puppet repo selector far enough down such that
all choices are hidden due to the use of Chosen. This can also occur if
the list of puppet repo choices is too long. Switching to a normal HTML
based selector allows the selector to appear a top of the details view
and prevents choices from being unreachable.