1002495 - Fixed a repo delete issue
Prior to this commit there was logic that said something along the lines
of 'Do not delete a RH repo unless its disabled'. Due to this when an org
was getting deleted, if there were any 'enabled' redhat repos in the org
there would be dangling repositories with nil references.
This commit adds an additional clause that says, "Allow deletion of the
redhat repo if org is getting deleted".