Engine: First round of test enablement.
This is the first in a series of commits that will enable our
test suite. The major changes this first commit brings are:

Addition of a test and spec task that can be run from Foreman proper.
To run the test suite, ensure the engine is loaded by Foreman and run either
'rake test:katello:test' or 'rake test:katello:spec' to run test unit or specs
respectively.

Conversion of Rspec test suite over to Minitest::Spec. The first file
changed was spec/models/activation_key_spec. This starts down a path
to unify on a single test framework - Minitest and bring our test suite
into congruence. The biggest changes are around the use of Mocha to handle
complex mocking and changing from 'should' expectations to 'must' and 'wont'.

Clean up of fixtures to reflect the table names they are associated with. Further,
each fixture must explicitly declare the class name associated in the katello_test_helper.
Automatic fixture label resolution wasn't working, thus opted for the use of Fixtures.identify.