Engine: Dashboard: fix error related to unserialized preferences
With katello running as an engine, we've seen multiple cases
where an error occurs when loading the dashboard.  It appears
that under certain circumstances the preferences are not being
unserialized.  This change will force unserialization, when
necessary.

When it occurs, the stack for the error looks similar to:

no member 'dashboard' in struct

activesupport (3.2.14) lib/active_support/core_ext/object/try.rb:36:in `[]'
activesupport (3.2.14) lib/active_support/core_ext/object/try.rb:36:in `try'
/home/bbucking/github/katello/app/lib/katello/dashboard/layout.rb:41:in `setup_layout'
/home/bbucking/github/katello/app/lib/katello/dashboard/layout.rb:37:in `initialize'
/home/bbucking/github/katello/app/helpers/katello/dashboard_helper.rb:188:in `new'
/home/bbucking/github/katello/app/helpers/katello/dashboard_helper.rb:188:in `dashboard_layout'
/home/bbucking/github/katello/app/helpers/katello/dashboard_helper.rb:193:in `render_dashboard'
/home/bbucking/github/katello/app/views/katello/dashboard/index.html.haml:14:in

A similar issue was reported in:
http://stackoverflow.com/questions/15588643/rails-3-unserialize-returns-only-activerecordattributemethodsserialization