Allow UUID to be passed through candlepin
Previously, consumer UUIDs were not being passed through to candlepin on
consumer create calls. Candlepin would create a UUID and pass it back to
katello.

This patch sends the consumer UUID through to candlepin. This is usually set to
null, which results in candlepin choosing the UUID. However, a UUID can now be
optionally supplied.