Engine: Enabling Jenkins task for outputting test results in CI environments.
This commit includes changes to how the custom test runner is handled.
Previously the runner was explicitly required and set in the test helper,
but due to how the ci_reporter gem sets the test loader as an environment
variable we needed to override that. The test runner now inherits from
the ci_reporter test runner and is set as the test loader in the test.rake
file. This effectively binds our test files to needing to be run via the
rake task. Note also that the spec and test tasks have been combined into
one to rule them all.