1002495 - Fix related to org delete
An error related to modifying a new record was raised when deleting an organization.
During an org delete the provider gets deleted which inturn tries to
delete the tasks associated to it. The task had dependent => :nullify
set to true, which made the task update the provider marked for
deletion.
This fix removes the dependent destroy on task status when a provider is
deleted, since organization removes it. Also redhat provider is the only
kind with a task id associated to it and the redhat provider can be only
deleted when the org is deleted.