Engine: UI: adding back the org aspect of a user's session
The code for the katello user's session was previously removed, since
the sessions (e.g. create/destroy) will be managed by the core app;
however, the session also included the ability to set the user's current
org.

Since orgs are currently handled differently between the core
app and the katello engine, this commit will add back the management
of the user's current katello org.  The goal in the future will be
to have a single org representation.