Engine: models: include_root_in_json=false for all katello models
This change is to ensure that the root node (e.g. 'user', 'system'...etc)
is not included in results generated via 'as_json'.

This change will primarily impact the v1 APIs which return data
via as_json; however, it will not impact the v2 APIs that are
generted via rabl.

One of the reasons this change is necessary is to ensure compatability
with the subscription-manager which does not expect to receive named
objects.