Engine: only create pulp_server if remote_id exists
The remote_id will almost always exist; however, this change
is to address a single scenario where we know that it will not.
When the katello engine is initially included in to the foreman
core and the user runs db:migrate, this logic will get hit;
however, it is prior to the migration that adds the remote_id.