Engine: Organization: updates so that katello org is an extension of the core's org
This commit contains changes so that the katello organization
is now an extension of the core (foreman) organization.  The
result will be a single organization (taxonomies) schema.

Based on discussion with the team, since we are early on
with the engine work, we are purposely updating the existing
migration to remove 'katello_organizations' vs trying to
migrate data from katello_organizations to taxonomies.
This means users will need to katello:reset and
then db:migrate / db:seed.