Engine: System Groups: update initial migrations for nutupane impl
When the system groups nutupane was created, a couple of modifications
were made to the schema.  Those were tracked in the 'remove' migrations
that are being deleted with this commit.

We are currently in the process of rebasing the engine branch
off of devel; therefore, to minimize migrations, this commit
rolls those changes in to the initial 'katello_tables' migration.