Allows development gems to be declared that can be ignored in production mode by bundler.
When using a gemspec to define all dependencies, if the gem is included
using 'gem' in Bundler all gems marked as 'add_dependency' will be included
when 'bundle install' is ran. However, this means that if development
dependencies are declared in this same manner, they will be required by
bundler in production. This can be avoided by including the gem using 'gemspec'.
Using the 'gemspec' notation places all gems marked with 'add_development_dependency'
into a development bundler group. Thus, in development a developer can run
'bundle install' and get all dependencies needed to develop. In production,
'bundle install --without development' can be run and the gems marked with
'add_development_dependency' will be excluded from the bundle.