Engine: API: fix issue with resource_list for api_version in resource
The API resources were recently updated to optionally include
the :api_version.  This broke the logic that returned
the list of resources in the resource_list api.  As a result,
subscription-manager commands (e.g. register, environments,
orgs) will not work.

This commit will ignore the :api_version in the RAILS
resources  when determing the list to return
as wellas omit it from the results sent to the
client.

For example, without this change, the API returns:
{'entitlements': '/api/entitlements/', 'status': '/api/status/', 'packages': '/api/packages/', 'consumers': '/api/consumers/'}

With this change, it will return:
{'status': '/api/status/', 'organizations': '/api/organizations/', 'roles': '/api/roles/', 'tasks': '/api/tasks/', 'sync_plans': '/api/sync_plans/', 'users': '/api/users/', 'distributors': '/api/distributors/', 'providers': '/api/providers/', 'entitlements': '/api/entitlements/', 'gpg_keys': '/api/gpg_keys/', 'repositories': '/api/repositories/', 'environments': '/api/environments/', 'content_views': '/api/content_views/', 'system_groups': '/api/system_groups/', 'products': '/api/products/', 'systems': '/api/systems/', 'changesets': '/api/changesets/', 'activation_keys': '/api/activation_keys/', 'content_view_definitions': '/api/content_view_definitions/', 'packages': '/api/packages/', 'consumers': '/api/consumers/'}