Update and implement Provider actions for API v2
All v1 code has been brought over to the v2 api.
The show route shows all providers, so it is up to the caller to
actually filter the results.

Modifications and additions include:

  - Merges initial work done in bkearney/provider-command
  - Alphabetizes and groups together routes
  - Modifies import_manifest action to handle uploading zip files
  - Fixes where Product not responding to last_sync breaks Provider
  - Adds tests for CRUD actions
  - Adds a provider_params method to ProviderController
  - Adds the api_resources, and exposes child nodes for products and
    repositories
  - Updates routes for providers controller
  - Adds top-level provider routes for CRUD in addition to index
  - Moves member routes and alphabetize them
  - Changes respond_for_index to respond

This change also removes resource description block, since This
functionality is now in the base class: Api::V2::ApiController

Routes and actions related to Products were left in, since work
in those areas of the code is being completed in another branch.

Changes based on feed in PR #3533 include:

- Remove trailing whitespace and extra line breaks
- Make rubocop happy
- Replace string literals with constants for provider_type
- Update description for index api
- Fix whitespace
- Make parameter order consistent
- Update organizations routes for providers to just index
- Put query logic in products action on one line
- Change @provider to @resource per convention
- Add param rules for ProvidersController
- Convert array literal in provider test setup to %w