Fixes issue where running 'rake katello:reset' wouldn't always reset the Candlepin database correctly.
Prior to this patch, the reset script assumed the existence of 'tomcat6'
which is not always the case. On most Fedora's, tomcat is under 'tomcat'
which resulted in the service never being stopped and Candlepin database
not being properly reset. This updates the task to check which tomcat
is installed and use it.