Changes required to have katello:reset work with a dev environment.
I first hit namespace issues, which is fixed in the http_resource
class. I then was getting wierd candlepin errors. This was due to
the keystore (which this does not replace) in /etc/tomcat/keystore
having been generated with a password which was not stored in
/etc/katello/keystore password. This resulted in a

javax.crypto.BadPaddingException: Given final block not properly padded

exception.

Finally, the cpsetup actually restarts tomcat, so I removed the extra
commands which are not needed.