Content Views Rework: removing the puppet module filter
This commit removes the main code for puppet module filters.
With the rework, puppet modules will become a direct association
with the content view (and not through a traditional filter).

Note: this commit does not address changes to the actual publishing
of a content view, which is currently being worked on.