Content View Rework: update content view puppet modules to always include name/author
A content view puppet module may consist of:
- name
- author
- uuid (optional)

With this change:
- name + author => indicates to apply the latest version of the module
  during content view publish
- uuid + name + author => indicates to apply the specific version
  associated with the uuid to the content view during publish

From a user point of view, this means the user can simply:
- on create or update, provide name+author or uuid

If the user provides uuid, the application will look up the puppet
module in pulp and assign the name + author (i.e. the user doesn't need
to provide it).