Content View Rework: update foreman puppet import
With the content view rework, we no longer associate puppet repositories
with a content view.  Instead, we associate puppet modules (which
could be from any puppet repo in the library for the org).

This commit takes that in to account when determining if there
are puppet modules to import in to foreman.