fixes #4726 - update hostgroup activation keys to support katello v2 apis
When a user selects a lifecycle environment & content view,
a request is sent to the server to retrieve the activation
keys associated with the selected env+view.  This allows the user
to choose from those via the Activation Keys pane.

This commit is to fix the behavior which previously
performed multiple requests to the server to retrieve
environment, content view and then keys.  This alters
that behavior to perform a single request for the keys.