fixes #4773 - Content Views: fix issue which did not allow for multiple errata ids per filter
The erratum filter rule 'types' is a serialized array.  The scope
that was placed on the model was incorrectly looking for 'nil';
however, with a serialized type, we really need to fetch the records
to look at the value of 'types', as the value for an existing
rule will not be 'nil', even when types is not specified.