fixes #4776 - session - extend session timeout when api controllers are hit
The Katello UI uses it's API controllers heavily; however,
the Foreman core only extends the session when a UI controller
is hit.  As a result, a user that is primarily using the Katello
UI can experience more frequent timeouts, if they stay on the same
page to perform several actions.

This commit will update the base Katello API controller to extend
the session without impacting the behavior of the core.