Fixes #4701 - unable to search unicode strings in elasticsearch
Asciifolding filter maps all unicode characters to their ascii equivalents. The mapping is destructive and prevents the original forms from being looked up. This behaviour is going to be fixed in elasticsearch 1.1 with option 'preserve_original'. I'm disabling the filter until the new elasticsearch is out and we can migrate to it.

See docs for more details:
http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/analysis-asciifolding-tokenfilter.html