Refs #4883: Disables SMAPs from showing up in the UI menu.
With the recent UI updates to Nutupane, and styling introductions of
Patternfly and Boostrap 3 (post enginification) as well as the backend
updates to the code base in favor of DynFlow, the Subscription Manager
Applications page is essentially non-functional. This temporary solution
is to remove the menu item for Subscription Manager Applications so as
not to give users a false sense of an existing feature.