fixes #5013 - host groups - fix retrieval of activation keys
This commit fixes the behavior of hostgroup create/edit,
where the user selectes a lifecycle environment and
content view and in turn the list of available activation
keys is return from the server.

The main issue prior to this commit was that the
logic was using the foreman puppet environment id
where it should have been using the katello
content view id.