fixes 5039 - update bulk actions before_filters to enable proper authenticate
Prior to this commit the system bulk actions for 'errata' would
fail with a 403 forbidden indicating that the user was not set.
The root cause is that User.current is currently set as
part of 'authenticate' (which is performed by the 'authorize' filter);
however, authorize was being executed after some of the filters that
required access to User.current.

This commit reorders the filters so that User.current
will be set before it is needed by other filters.