Fixes #5098: Table select all checkboxes no longer trigger one another.
Previously, the select all functionality was tied to the assumption that a
table variable existed attached to the scope; however, this use of a table
variable reflected that created by an outside controller and not the internal
scope of the directive.

Changes the handling of select all internally within alch table by
introducing a private variable that belongs to each instance of alch-table.
Methods were introduced to wrap the internal variable to enforce the internal
nature of the variable.