Allows updates of Products by ID only.
Currently one has to provide both the Product's ID and NAME to update
any of the available fields. This change should allow us to update a
product by using only its ID. Fixes #5133 and BZ 1085536.

Note that since a product label cannot be updated, I removed it from the
param_group set and manually added it to the create method (optional).